Abstracts chart functions for use in 'Predictions' page
